refactor: migrate to cosmos-sdk 0.42.2
all tests successful

+++ x/bitcoin/abci.go
@@ -39,7 +39,7 @@ func EndBlocker(ctx sdk.Context, req abci.RequestEndBlock, k types.BTCKeeper, si
 
 	var sigs []btcec.Signature
 	for i, in := range outpointsToSign {
-		hash, err := txscript.CalcWitnessSigHash(in.RedeemScript, txscript.NewTxSigHashes(tx), txscript.SigHashAll, tx, i, in.Amount)
+		hash, err := txscript.CalcWitnessSigHash(in.RedeemScript, txscript.NewTxSigHashes(tx), txscript.SigHashAll, tx, i, int64(in.Amount))
 		if err != nil {
 			k.Logger(ctx).Error(fmt.Sprintf("could not calculate sig hash for outpoint %s", in.OutPoint))
 			return nil
